feat: GuestScan Component
